+++ libraries/QualitySample/src/QualitySample.cpp
