+++ libraries/QualitySample/tests/QualitySampleTest/Makefile
@@ -1,0 +1,10 @@
+# See https://github.com/bxparks/EpoxyDuino for documentation about using
+# EpoxyDuino to compile and run AUnit tests natively on Linux or MacOS.
+
+APP_NAME := QualitySampleTest
+ARDUINO_LIBS := QualitySample AUnit
+include ../../../../EpoxyDuino/EpoxyDuino.mk
+#ARDUINO_LIB_DIRS := \
+#/mnt/c/Users/elauber/Documents/Arduino/airgradient_mqtt/src/Sampling/QualitySample.h
+#$(arduino_ide_dir)/airgradient_mqtt/
+#../../src/Sampling/QualitySample.h
