--- tests/README.md
@@ -1,4 +1,0 @@
-# AceButtonTest
-
-These unit tests depend on
-[AUnit](https://github.com/bxparks/AUnit).
