--- tests/Sampling/Makefile
@@ -1,7 +1,0 @@
-# See https://github.com/bxparks/EpoxyDuino for documentation about using
-# EpoxyDuino to compile and run AUnit tests natively on Linux or MacOS.
-
-APP_NAME := QualityTestsTest
-# may need to change library stuff here
-ARDUINO_LIBS := AUnit Sampling
-include ../../../EpoxyDuino/EpoxyDuino.mk
